Folder restructure, extension handling
