--- CountriesListApp/AppModules/View/DetailRowView.swift
@@ -1,38 +1,0 @@
-//
-//  DetailRowView.swift
-//  CountriesListApp
-//
-//  Created by Naval Hasan on 08/06/25.
-//
-
-import SwiftUI
-
-struct DetailRowView: View {
-    let title: String
-    let value: String
-    let icon: String
-    
-    var body: some View {
-        HStack {
-            Image(systemName: icon)
-                .foregroundColor(.blue)
-                .frame(width: 24)
-            
-            VStack(alignment: .leading, spacing: 2) {
-                Text(title)
-                    .font(.subheadline)
-                    .foregroundColor(.secondary)
-                
-                Text(value)
-                    .font(.body)
-                    .fontWeight(.medium)
-            }
-            
-            Spacer()
-        }
-        .padding(.vertical, 8)
-        .padding(.horizontal, 12)
-        .background(Color(.systemGray6))
-        .cornerRadius(8)
-    }
-}
